SecRL/requirements.txt
tiktoken
networkx
pandas
sqlalchemy
gymnasium
numpy
rich
gymnasium
pyautogen==0.2.35
azure-identity
azure-monitor-query
mysql-connector-python
python-Levenshtein
matplotlib
azure-ai-ml
azure-ai-inference
bs4
datasets
langchain-huggingface
langchain-openai